prov/psm: move fabric implementation into a separate source file
