+++ prov/psm2/src/psmx_fabric.c
@@ -1,0 +1,1 @@
+../../psm/src/psmx_fabric.c
